[#17] Make triangles only compute containing plane once
- Move PlaneContaining() to Triangle files instead of Plane files
- Compute PlaneContaining() for each triangle as its created
- Remove an extra .Normalized() from PlaneContaining()
